feat: 封装对话对象，添加controller方法
将XiaozhiAgent中的chat方法的memoryId类型修改为Long

+++ src/main/java/com/ai/xiaozhi/assistant/XiaozhiAgent.java
@@ -13,5 +13,5 @@ import dev.langchain4j.service.spring.AiServiceWiringMode;
 )
 public interface XiaozhiAgent {
   @SystemMessage(fromResource = "XiaozhiSystemPrompt.txt")
-  String chat(@MemoryId int memoryId, @UserMessage String userMessage);
+  String chat(@MemoryId Long memoryId, @UserMessage String userMessage);
 }
